requirements.txt
streamlit
pyserial
pandas
matplotlib
numpy